backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.17
httpx==0.27.2
python-dotenv==1.0.1
sqlalchemy>=2.0
psycopg2-binary>=2.9
resend>=2.0.0
